Extract ReleaseDescriptionEditor component
Move release description editor, upload processing and save
logic into a dedicated component. Remove editor state and save
handlers from the release page and update imports. Tidy a few
imports and className whitespace issues.
